Lamp Added and Fixed Bug
Added a lamp prefab using the previous torch Object. Need to make it so it just turns on the light because right now it has the entire object not active then sets it active. Also fixed a bug where the Activated boolean would turn on and off in the backend.