Made a Designer Friendly Folder to Create the Dungeon
I made a folder in the Vertical Slice folder called SMM-Dungeon which contains all the prefabs and dungeon setup necessary for a designer to design the dungeon. They definitely need help though.
I made a template to use when making a new scene. Tell them to copy the SMM Dungeon scene or use the template to make a new one.