Switches and doors now have a reference to the switches/doors referencing them
Added an OnValidate method to each switch and door to add themselves to an array on all object that they reference. This is so you can see all the references to each switch that target a door on the door itself, while also seeing all doors that the switch targets on the switch. Might be buggy, seems kinda unpredictable. Not sure if worth but kinda cool.