allowed dialogue choices to expand when selected
The dialogue box now expands to fit the size of the text when selected, and will go back to normal when deselected.

Small Bug:
when a choice is first loaded into the screen, it will fit to the text size and if it is the only option, you can't deselect it. This results in the cutscenes where you only say "Next" have a slightly smaller box around the "Next" option.