Magic Hand beta base
Added the magic hand; the hand can be spawned with the C key, and despawned early with the same key, with the cooldown being reduced if you do so. However, functionality is currently incomplete and mildly bugged; the hand can grab objects, which go into it's inventory, and I have yet to code in the ability to confuse enemies by touching them with the hand. There also seems to be an issue regarding collision; the hand and player can collide with one another and push eachother around, and I'm not familiar enough with our current collision system to fix. Also, I think I might have broken the player prefab, or at least the player in the scene I was testing in.