Implemented sound effects for sight beast
The functions for calling sounds from the sight beast are now functional! Additionally, the monster now plays the idle sound effect when walking around, and the alert sound effect when it begins a chase. Currently no sound effect plays while chasing; I was unsure if it was supposed to, and if so which sound effect should play.